+++ packages/typespec-ts/test/azureIntegration/bodyParameters.spec.ts
@@ -26,7 +26,8 @@ describe("BodyOptionalityClient Rest Client", () => {
       .post({
         body: {
           name: "foo"
-        }
+        },
+        contentType: "application/json"
       });
     assert.strictEqual(result.status, "204");
   });